requirements.txt
newspaper3k
requests
openai
pandas 
numpy 
matplotlib 
seaborn
scipy
scikit-learn
python-dotenv
